fix error response code

+++ src/controllers/record.controller.js
@@ -14,7 +14,10 @@ const { checkErrors } = require("../services/error.service");
 
 const getRecords = async (req, res) => {
   try {
-    checkErrors(validationResult(req), res);
+    if (checkErrors(validationResult(req)) === 401) {
+      res.status(401).send({ msg: "Unauthorized" });
+      return;
+    }
 
     const records = await getAllRecords(
       req.query.pageSize,
@@ -40,7 +43,10 @@ const getRecords = async (req, res) => {
 
 const getRecord = async (req, res) => {
   try {
-    checkErrors(validationResult(req), res);
+    if (checkErrors(validationResult(req)) === 401) {
+      res.status(401).send({ msg: "Unauthorized" });
+      return;
+    }
 
     const record = await getRecordById(req.params.id);
     console.log(record);
@@ -55,7 +61,10 @@ const getRecord = async (req, res) => {
 
 const postRecord = async (req, res) => {
   try {
-    checkErrors(validationResult(req), res);
+    if (checkErrors(validationResult(req)) === 401) {
+      res.status(401).send({ msg: "Unauthorized" });
+      return;
+    }
 
     if (!(await checkDoctorExists(req.body.doctor))) {
       return res.status(404).json({ errors: "Doctor not found" });
@@ -72,7 +81,10 @@ const postRecord = async (req, res) => {
 
 const patchRecord = async (req, res) => {
   try {
-    checkErrors(validationResult(req), res);
+    if (checkErrors(validationResult(req)) === 401) {
+      res.status(401).send({ msg: "Unauthorized" });
+      return;
+    }
 
     if (!(await checkDoctorExists(req.body.doctor))) {
       return res.status(404).json({ errors: "Doctor not found" });
@@ -89,7 +101,10 @@ const patchRecord = async (req, res) => {
 
 const deleteRecords = async (req, res) => {
   try {
-    checkErrors(validationResult(req), res);
+    if (checkErrors(validationResult(req)) === 401) {
+      res.status(401).send({ msg: "Unauthorized" });
+      return;
+    }
 
     const deletedRecords = await deleteAllRecords();
 
@@ -103,7 +118,10 @@ const deleteRecords = async (req, res) => {
 
 const deleteRecord = async (req, res) => {
   try {
-    checkErrors(validationResult(req), res);
+    if (checkErrors(validationResult(req)) === 401) {
+      res.status(401).send({ msg: "Unauthorized" });
+      return;
+    }
 
     const deletedRecord = await deleteRecordById(req.params.id);
